Enable save upload with file as body
There's two ways to upload a file. One is through the use of
multipart/form-data, where the file is one part and the metadata is
encoded as JSON in another part. This is how the UI uploads files.

The other way to upload a file is to use HTTP headers to encode metadata
and use the body as the file contents. This is how the CLI client works.
It's simpler a call and could facilitate efficiency for clients that
are able to stream requests from disk.

The second functionality was mistakenly not ported over, so this commit
enables it.

The cap on file size is hoisted up to the reverse proxy.