Rename internal modules and add contributor guide
admin-shared was renamed to applib to better reflect that the library
wasn't only used for admin purposes.

eu4gamejs was renamed to applib-node to better reflect that the module
houses the N-API bindings. If future games are added to the backend,
they would be exposed to the app backend here so the name should be game
agnostic.

All packages are marked as private to prevent accidental uploads. This
is in preparation in case it is desirable to grow the monorepo to
include published save libraries like eu4save, hoi4save, etc.

Add a dash of contributor docs to the readme. As sections grow they can
be split off into their own dedicated readme and referenced from the
TOC.