Produce minimal diffs from db shift
When there is a change in save parsing logic, the database will start to
drift and contain out of date or erroneous information. Scenarios most
susceptible to dift:

- Previously uploaded saves qualify for the new achievement added
- Determining playthrough id algorithm changes
- Incorrect logic for extracting info like player names or days since 1444

This commit compares the current state of the database as a csv and
reparses all the saves and produces the minimum amount of JSON that
needs to be applied to database to make it current.