Enable toggle tag filter in info drawer
When the user clicks on the "Show only X on the map" button in the info
drawer, we show only that tag on the map. It seems reasonable that if
the user were to click the button again, it would reset to the map mode
at the time when the user opened the info drawer.