Fix standalone map asset references
When the map repo was integrated into the main repo, some of the map
assets were moved around so that they could be easily positioned in git
lfs to not disturb the weight of the main code and could be easily wiped
as needed.

Unfortunately, I forgot to update the standalone map asset references,
so one couldn't run the map repo anymore.

This commit fixes that issue

Closes #36