Move clear select items 'X' from right to left
Moving the select items 'X' from right to left means that the 'X' is now
in the same position relative to whatever it's bordering. This means
that one can more easily spam click to remove nations (though, depending
on how many countries one removes, it may be better to start with a
smaller base set of countries, like players and explicitly include
countries).