Update social image, title, and description
Based on feedback from https://socialsharepreview.com/?url=https://pdx.tools
I updated the image to be a slightly different view. It was pointed out
that pie charts and bar charts are probably easier to distinguish
compared to line charts when the image is small -- like it could be when
shared.

Title and description were slightly updated to match recommendations.