Fix initial map render omitting stripes and highlights
When two consecutive saves are loaded, we preserve the map mode between
the two to ease a sort of comparison. However, we must always calculate
political borders on save load to allow for "paint country borders"
option to work regardless of map mode.

An issue of missing stripes and highlights occurred due to the first
render using this faux color calculation.

Closes #48 #37