Configure and update sentry to latest
As part of the migration to pdx.tools, the sentry environment variable
stopped being [exposed to the browser][1]. This cut the app off from
error reporting, leaving us a bit blind to user errors. The fix is to
export `SENTRY_DSN` as `NEXT_PUBLIC_SENTRY_DSN` so that the browser sees
it.

This commit also updates sentry's library to the latest version.

This commit also removes performance trace sampling so that we won't
have to worry about exceeding our quota again.

[1]: https://nextjs.org/docs/basic-features/environment-variables#exposing-environment-variables-to-the-browser