Add dev container support
Right now setting up a dev environment is a little cumbersome as one needs the
right flavor Linux (debian based) and the right dependency versions. One must
also install several dependencies globally. This is no fun for not only
newcomers who may not have a bunch of package management to draw upon but it
is also an inconvenience to me. I may need to RMA a motherboard soon so
anything that will help me easily transfer the development setup is huge.

The fix is to introduce [devcontainer][0] support. Yeah it's vscode and github
codespaces specific, but it seems to work flawlessly on both windows and
linux, so having this is already a ton better than no dev environment
assistance.

This PR also removes the dependency on the `brotli` command in favor of
using the rust brotli dependency that is already included in other
modules

[0]: https://code.visualstudio.com/docs/remote/containers