Fix permission issue for saves on google drive
On my phone I was receiving the following error when I selected a save
in google drive:

> The requested file could not be read, typically due to permission
> problems that have occurred after a reference to a file was acquired.

The fix is to read the save before transferring it to a web worker. Not
sure of the technical explanation of this or where it is documented, but
it worked, so I won't question it.

After this fix, I can confirm that I can load saves from google drive
and microsoft one drive from my phone and so closes #78