Decouple binary tokens from Wasm and support CK3 1.5
First and foremost this updates CK3 support to melt 1.5 save games.

But the more interesting update is that the binary tokens are no longer
compiled into the Wasm. Now when a game module is instantiated, instead
of fetching just the wasm payload, the tokens are fetched too. This was
done for several reasons:

- As a form of code splitting. Now if only one of the payloads (the wasm
  or tokens) change, the other doesn't need to be redundantly fetched
  too.
- Compile times. Previously it would take 6 minutes (2 to compile, 4 to
  optimize) the HOI4 module. Now it is much shorter.
- Preparation for a future where token incompatibility is real and we
  may need to fetch a different token list depending on the save version

Performance thus far seems unaffected.