Fix recent tag switch exclusions from map filter
The country filter tries to be smart and not include tags that have
never existed in the game in the filter. It does this check by examining
province history and if there is no record of a tag owning a province,
that tag is excluded.

The problem is when a tag switch occurs the province history entries
aren't ported over to the new tag, so if the recent tag switch doesn't
acquire any more provinces, it will be as if it doesn't exist.

The fix is to include current province owners in addition to province
history in the country filter