Remove EU4 achievement A Manchurian Candidate
A manchurian candidate is currently not able to detect achievements
correctly as the manchu formation will rewrite the original country's
primary culture history from `manchu` to `manchu_new` which won't match
the achievement conditions.

It seems like the fix would be to examine the game files for the true
initial primary culture. Until then, we will disable the achievement.