Allow recording map timelapse into a video
This feature allows users to capture a timelapse into a video recording
to download and share. This implemented at a high level by using a
[canvas's `captureStream`][2] and piping that to a [`MediaRecorder`][3].

Users may choose what resolution to render the map as well as the how
fast the timelapse progresses.

Conceptually simple, but several problems arose.

Capturing the map's canvas omits UI elements like the current date in a
timelapse, something that seems critical when evaluating a timelapse. To
embed a date in the recording, we'd need to either send text into the
map or we could copy the map's canvas into a separate canvas that uses
the 2d api, which has better support for text and we could record that
one instead. While copying around map canvas data seems inefficient, it
was deemed better than needing to figure how to embed rapidly updating
text into WebGL. To support the copy use case the webgl map exposes a
callback (`onCommit`) right after it has rendered so the recording
canvas can copy it hot off the press.

Since users will want the option to record the entire map, the recording
logic communicates to the canvas that it'll be assuming control of the
size of the canvas. Seeing the canvas expand beyond one's viewport may
be unsightly, but it was the most intuitive way to get the resolution
necessary for the recording. The good news is that we restore the map
state when the recording is done so the user's focus point and scale
will not have changed.

The hardest part about the implemenation is all the coordination between
updating the map, progressing the timelapse, and making sure that
recording captures both the start and tail end of the timelapse.

For instance, there are some variations in browser implementations of a
canvas's capture stream and the media recorder. The biggest one is that
chrome would cut several seconds out of a recording especially when
there was a lack of data (like when using a freeze frame at the end).
The solution is extremely hacky: when the recording stops, repeatedly
write a transparent rectangle until we the media recorder emit non-empty
data.

That same technique (writing transparent data) is used to synchronize
the start of the recording with the start of the timelapse.

Another variation is that Chrome's canvas stream needed to be updated
every 250ms or it wouldn't register any changes. This is why the minimum
intervals per second is 4.

The output of a recording is a webm file. While this format is the [
recommended approach for video playback][4], webm upload is not often
supported sites like reddit. The solution is that after the recording is
finished, transcode it into an mp4 file using [`ffmpeg.wasm`][5]. This
is a pretty extreme approach as ffmpeg is a large download (several
times larger than all the pdx.tools wasm bundles put together), and may
take several minutes to complete while using 100% CPU usage. But the end
result is worth it if users can take the download and immediately upload
it elsewhere with minimal hassle.

However I still view the mp4 export as a bit experimental as every few
recordings ffmpeg takes 10x as long to process the input and the output
is horribly choppy. I'm not sure what causes this, but I'm assuming
other users will experience this issue, so my recommendation to those
who constantly experience it is to accept the webm output and convert it
to mp4 through a 3rd party tool.

Needed to add two new headers: `Cross-Origin-Embedder-Policy:
require-corp` and `Cross-Origin-Opener-Policy: same-origin` in order to
get [ffmpeg.wasm to work][0] as it relies on `SharedArrayBuffer` and its
[security requirements][1]

The timelapse and recording work through `requestAnimationFrame` which
is paused when running in a background tab ([source][6]), so in order to
avoid corrupting the canvas capture, a warning is presented to the user
when they switch back to our tab (as one can't eagerly detect tab
switching in the browser beforehand), informing them of this
restriction. At least the MP4 transcoding can happen in the background.

[0]: https://github.com/ffmpegwasm/ffmpeg.wasm#installation
[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer#security_requirements
[2]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLCanvasElement/captureStream
[3]: https://developer.mozilla.org/en-US/docs/Web/API/MediaRecorder
[4]: https://developer.mozilla.org/en-US/docs/Web/Media/Formats/Video_codecs#recommendations_for_everyday_videos
[5]: https://github.com/ffmpegwasm/ffmpeg.wasm
[6]: https://developer.mozilla.org/en-US/docs/Web/API/window/requestAnimationFrame