Add export to csv to chart view
The visualizations and data tables in the chart view may not be in
someone's desired format, so this commit adds the ability to download
the data as a csv.

The main implementation oddity is that since the data download button
lives outside any calculations, it relies on the visualizations to set a
function to lazily compute the csv.