Splice token gap from lookup
All tokens observed have some sort of large gap (8,000 or so) in token
values that picks back up at 10,000.

This commit takes advantage of this observation to splice out this gap
so that we don't store and communicate so many empty strings for the
lookup. Benchmarks confirmed that performance hit is negligble. There is
a savings of around 100 KB uncompressed per token list, which isn't bad,
but these savings tend to vanish after brotli decoding.