Optimize political map timeline calculations
There's something about the province filters that slows down the map
calculation. Since the political timeline doesn't utilize the province
filters, we can skip that, which decreases time to calculate 2-3x

I also took the time to refactor the code a bit so that it should be
more readable and re-usable when it comes time for the religion map
mode.