Fix occassional poor quality and performance of mp4 transcode
For whatever reason, there are occassional recording runs that when
transcoding the webm to mp4 will cause ffmpeg to log errors about
duplicate frames. This effects both native and browser ffmpeg (ie: I
could take an exported webm file and see the same errors when manually
transcoding to mp4).

The transcoding process would be extremely slow, requiring minutes
whereas it would normally take seconds. The quality output would also be
horrendous and nothing like the webm.

The fix is to set vsync to passthrough as mentioned here:
https://superuser.com/questions/1308355/ffmpeg-webm-to-mp4-awful-video-quality#comment1944969_1308355

I'm not too well versed in ffmpeg flags but it seems like webm could
have a variable bit rate and that caused friction in the transcode.