Remove redis and fold leaderboard and API keys into Postgres
What does this mean for users:

- Uploads are no longer partially gated by requiring the save to crack
  an achievement leaderboard.
- API keys need to be regenerated.
- The `/eu4/achievements` page no longer shows the current record for
  each achievement

Redis has been with PDX Tools since day 1. The reason being that Redis
is tailored made for a leaderboard with its sorted sets; making it easy
to see if a save cracked the leaderboard and other operations.

It seems only natural to want to use the right tool for the job, so
Redis was put to work.

Later on API keys were added to Redis.

Why then remove Redis and fold the leaderboard and API keys into Postgres?

Ease of maintenance mostly. In our user interest survey, enriching
achievements lost in a head to head matchup with every other proposed
development task. And in the 2 years of running PDX Tools, the casual
leaderboard movement hasn't been embraced.

So it makes sense to de-emphasize achievements, but since I still enjoy
achievements, they aren't going away. I will continue maintaining them
if only to satisfy my desires, but I want them to be as easy to maintain
as possible. And running a separate database just for achievements is
not a worry free solution as there are some cross cutting concerns to
address:

- Deployment
- Configuration
- Authentication
- Logging
- Monitoring
- Backup
- Dependency management

I realized I could get rid of Redis if I added a score column to the
postgres database.

Then a bunch of code pertaining to the leaderboard could be deleted. As
consequence, the leaderboard is no longer checked prior to uploading a
save in order to spare the user's save slot.

Also I could imagine calculating the top save for each achievement via
Postgres to be prohibitively expensive, so this was dropped as well.

API keys are now in Postgres and require a regeneration.