Timelapse video switch from MediaRecorder to VideoEncoder
The current video output from timelapses can be a little janky:

 - The starting map may be incorrect
 - Recording may skip years if user's computer can't keep up with the
   selected frame rate.
 - Recording timelapse can only proceed at the selected framerate

The fundamental issue is that timelapse recordings worked by literally
recording the screen. And it improved impossible to coax the correct map
rendering at the correct time. We got it pretty good, but only through
lots of hacks (like continually rendering a start and end frame until we
data was encoded)

This commit changes the video paradigm drastically. Now a video is
directly created by individual frames where we can control the content,
timestamp, and duration of the frame.

Unfortunately, the required [`VideoEncoder`][0] has limited browser
support with only Chrome, Edge, and Opera really supporting the feature.
This represents 75% of pdx tools users.

Using a `VideoEncoder` required an additional library in order to mux it
into WebM and then ffmpeg is still required for an mp4 transcode.
Annoying, but at least the library is lightweight.

Of note: I tried using a vp9 codec, and even though Chrome supports vp9
for a `MediaRecorder`, it does not support it for `VideoEncoder`. I
can't say why.

[0]: https://caniuse.com/mdn-api_videoencoder