Update home page with image gallery
The current home page image is a single screenshot that is blurry on
high DPI monitors. This commit introduces a gallery hosting several
images that can be selected from. Each image has a title and a short
description. The images are all high resolution.

A religion timelapse video is also included in the gallery.

The gallery is information dense, so on mobile the image gallery is
rolled out and images that are composed of other images use only a
single image of their composition as a substitute.

Couple minor changes:

 - Country states has the true/false column substituted with a checkmark
 - Country rulers average stats round to 2 decimal places like other
   columns.