Support VP9 timelapses and define bitrate function
Closes #154

Adds in a bitrate that is a function of the canvas size with a minimum
of 100 kb/s and a max of 2,000 kb/s. The output seems good enough for
now. Ideally the bitrate could also be a function of if terrain is
overlaid or what map mode is selected. But we'll keep it simple for now.