Fix religious timelapse color calculation
At the save date the religious map mode looks at the province
controller's religion and the province's religion (consistent with
in-game). While in a timelapse, the province owner's religion is
compared with the province's religion.

This mismatch in logic can result in funky changes in a recorded
timelapse if an area is under occupation as the controller won't match
the owner.

To remedy the issue, the map payload will always denote the date is null
if the date is the date of the save. Since the timelapse always
transmits a date, it will use the province owner even at the end date.