Replace HTTPS dev environment with staging
Previously there was an HTTPS dev environment to test features that
required a [secure context][0]:

- Firefox's brotli decoder
- SharedArrayBuffer (timelapse)

I'm not sure when, but now browsers recognize HTTP over localhost as a
secure context so we can drop everything related to HTTPS.

And what's better is that the HTTPS dev environment has been replaced by
a staging environment where a local cloudflare instance is spun up. That
way, it is possible to test all the cloudflare logic without needing to
deploy it to the dev server.

[0]: https://developer.mozilla.org/en-US/docs/Web/Security/Secure_Contexts