Add save watcher initial implementation
Using the file system access api (mainly available on chrome), this
commit exposes the ability for the user to opt into having their save
file watched for changes, and when changes are detected -- the save is
reparsed and the UI updated (map included).

Right now it is tucked away under the religion tab.