Migrate from prisma and drizzle
I see the cool kids doing it, so obviously I have to do it.

I kid but I do want to start moving dependencies towards an edge
readiness.

I also notice that the spikes in CPU usage on the server and I'm not
sure what it is from, nothing is tipping me off, so I'm hoping this
helps.