Add architecture and design strategy document
The architecture was a bit outdated, so it was replaced with a diagram
that should make it easier to see how the components are related to each
other.

A design strategy document was added that highlights some of the PDX
Tools tenets and how past experiences relate to them. I think a lot of
cool things have been done and I don't want to forget them.