Add the most request flags file
Right now it is unused, but who knows, maybe it can be used to split out
hot sprites so they can load quicker, or maybe they can served at a
higher quality.

This file was taken from access logs before the transition to sprites,
and I don't want to lose this information.