Remove unused headers from the ffmpeg.wasm days
`Cross-Origin-Embedder-Policy` and `Cross-Origin-Opener-Policy` were
required when ffmpeg.wasm was used in order to get shared array buffers
to work.

For some reason, these headers were stripped from static assets (JS and
CSS) but not from pages on Cloudflare pages, which caused the JS
referenced in a web worker to fail to load due to a policy issue.