Support cloudflare pages next.js deployment (#247)
* Support cloudflare pages next.js deployment

Starting to get close to the "Fast Origin Transfer" limit on Vercel for
what I believe is the S3 proxy for save files. I don't believe it is the
bandwidth from Cloudflare (which underpins Vercel's edge runtime) to
Backblaze B2 (the S3 provider) as they are partners in the bandwidth
alliance, but rather the bandwidth from the edge runtime to the user.

Considering I'd like to keep a reverse proxy in front of S3 for the
flexibility, Vercel is starting to look limiting.

This commit changes the deployment to cloudflare pages. The following changes were needed:

- Remove `no-store` from the fetch cache as cloudflare doesn't support
  it. https://github.com/cloudflare/workerd/issues/698 Makes me question
  if Vercel's edge runtime actually runs on Cloudflare. `no-store` may
  not have been critical either.
- A patch was needed for `next-on-pages` to allow our split deployment
  with latent nodejs functions to be considered valid.
- `_routes.json` is added so that only api functions are counted as
  function executions. Otherwise users accessing docs or index.html (in
  a pure SPA) is counted against the Cloudflare function limit.