Add unique constraint on save hash column
Currently, there's application level code to check for uniqueness when
uploading a save. I believe I originally wrote the code this way to be
flexible (flexible in what way remains a mystery).

Well, it's been a few years and that flexibility isn't needed, so let's
use unique contraints so the DB can save us from the small race
condition.

I'm also on a bit of a kick to minimize the number of SQL statements
needed to execute