Reduce save SQL statements and ask for forgiveness
The `withSave` middleware would issue a `SELECT` (for sometimes more
data than needed) that wasn't marked in a transaction with `FOR UPDATE`
on patch and delete routes. Thus some race conditions were present.

This commit fixes this by using transactions and checking for permission
after the database has been mutated using the returned value from the
mutation. If the user didn't have permission, the database is rolled
back.