Move screenshot png creation off the main thread
The `toBlob` canvas function executes in the thread its on, which causes
any animation to be horribly stuttery.

This commit fixes the problem by creating an offscreen canvas and moving
the PNG creation to a worker thread.

Two years ago I dabbled in offscreen canvases and decided to ditch them
as the browser support wasn't there. Now offscreen canvas has been in
every browser since early 2023, so we should be good to go.

This commit also uses vite to run the map package as we now need a
proper bundle the web worker.