Render map webgl pipeline off the main thread
When investigating a better screenshot export user flow, I realized that
animations would stall due to the main thread being blocked.

`OffscreenCanvas` to the rescue. It's a pretty new API, but it has 96%
coverage from caniuse. That's good enough for me.

While smooth animation was the original goal, there's been a big uplift
in intialization: up to a 50% improvement.