Consolidate steam API routes
Instead of having a dev and prod steam routes that are defined
separately, conslidate the dev route into the prod route so any testing
or development also tests the db logic.

For instance, this commit also introduces a sql logic to know if a row
was inserted or updated when a verifies with steam. I don't want to
duplicate that logic to the dev route, so I consolidated them.