Add estimated land and naval morale
These are estimates. The land morale can be off due to the land
maintenance slider (which we can't unconditionally utilize as the slider
could have been changed before the month tick).

Naval morale is a bit more complicated with flagships giving a potential
fleet morale bonus too.

These are close enough to still be useful.