Add bumpalo serde support
Bumpalo serde extends serde's deserialization capabilities to work with
bumpalo. It provides an ArenaDeserialize trait, which is similar to
serde's Deserialize, but allocates data into a provided bump allocator
instead of using the global heap allocator.

Bump allocators and deserialization are a natural fit:

- A single drop for the deserialized model, means that no matter how
  many strings a model may have, it will always be constant time to free
  the data.
- Bumpalo allows one to pre-allocate the capacity expected for
  deserialization
- Amortizes allocation costs of many fields that require allocations
- Minimizes the volatility of the system allocator. Especially well
  suited for Wasm where one may have minimal control of the underyling
  allocator.
- Allows for efficient views into the underlying data without requiring
  the entire input to be kept around unlike borrowed deserialization
- Less brittle in performance than serde's `deserialize_in_place`.

This crate is highly experimental and is not published on crates.io, but can be once stabilized.