Add overlay middleware and GPU-rendered save date banner
The first layer I experimented with is imprinting the save date on EU5
saves when generating screenshots with eu5native. This is more of a
proof of concept, and indicates that it would be possible to replace the
EU4 og image preview of spinning up an actual browser with something
that is in process.

This layer renders the save date using a pixel font that is rasterized
onto a background and sampled in the shader. It might not be the
prettiest font, but it embodies pragmatic minimalism.