Update leaderboard podium to include country information
The leaderboard information currently omits the country the run started
with, as well as the country the run finished with -- two pieces of
information that is included in the achievement table below.

This commit fixes this bit of asymmetry by including a small country
flag on the podium. If the started and finished countries are different,
then both flags are displayed with an arrow going between them.