Convert compute shader to fragment shader
This seems to cut down on the tearing in firefox as we are now drawing
right to the surface instead of having a second draw phase for taking
the compute shader output and drawing to the surface.

And the transition to a fragment shader makes sense, as at the end of
the day, whether we are rendering to some surface or for a screenshot we
are ultimately rendering pixels to some target. There is no need to use
a compute shader.

This PR also splits up the Renderer object so that something low level
isn't holding onto the GPU connection, the pipelines, and the simulation
data -- and instead split them up.